+++ app/homepage/TodoLists.tsx
@@ -1,4 +1,4 @@
-import type { HomePageTodoListDto } from "shared/client";
+import type { HomePageDto } from "shared/client";
 
 import React from "react";
 import { isEmpty } from "fp-ts/Array";
@@ -6,27 +6,28 @@ import { EmptyMessage } from "front/ui/EmptyMessage";
 import { PageTitle } from "front/ui/PageTitle";
 import { TodoListItem } from "front/homepage/TodoListItem";
 import { AddTodoListForm } from "front/homepage/AddTodoListForm";
+import { useLoaderData } from "remix";
 
-interface TodoListsProps {
-  todoLists: HomePageTodoListDto[];
-}
+export const TodoLists = () => {
+  const { todoLists } = useLoaderData<HomePageDto>();
 
-export const TodoLists = ({ todoLists }: TodoListsProps) => (
-  <section className="space-y-10">
-    <PageTitle>Welcome, these are your todo lists</PageTitle>
+  return (
+    <section className="space-y-10">
+      <PageTitle>Welcome, these are your todo lists</PageTitle>
 
-    <AddTodoListForm />
+      <AddTodoListForm />
 
-    {isEmpty(todoLists) ? (
-      <EmptyMessage>No todo list has been added yet!</EmptyMessage>
-    ) : (
-      <ol className="space-y-2">
-        {todoLists.map((todoList) => (
-          <li key={todoList.id}>
-            <TodoListItem todoList={todoList} />
-          </li>
-        ))}
-      </ol>
-    )}
-  </section>
-);
+      {isEmpty(todoLists) ? (
+        <EmptyMessage>No todo list has been added yet!</EmptyMessage>
+      ) : (
+        <ol className="space-y-2">
+          {todoLists.map((todoList) => (
+            <li key={todoList.id}>
+              <TodoListItem todoList={todoList} />
+            </li>
+          ))}
+        </ol>
+      )}
+    </section>
+  );
+};
