+++ app/routes/forgot-password.tsx
@@ -1,0 +1,3 @@
+export default function ForgotPassword() {
+  return <div>coucou</div>;
+}
