+++ app/routes/welcome.tsx
@@ -1,19 +1,9 @@
 import type { MetaFunction } from "remix";
-import { PageTitle } from "front/ui/PageTitle";
+
+import { Welcome } from "front/welcome/Welcome";
 
 export const meta: MetaFunction = () => ({
   title: "TLM | Welcome",
 });
 
-export default function Welcome() {
-  return (
-    <div className="my-6 space-y-6 text-center">
-      <PageTitle>You're almost there...</PageTitle>
-
-      <p>
-        A email containing a verification link has been sent to your address.
-        You'll be ready when your account will be verified!
-      </p>
-    </div>
-  );
-}
+export default Welcome;
