+++ app/verify-account/useVerifyAccount.tsx
@@ -1,0 +1,14 @@
+import { useNavigate } from "remix";
+import { useEffect } from "react";
+
+export function useVerifyAccount() {
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    const timeout = setTimeout(() => {
+      navigate("/");
+    }, 5000);
+
+    return () => clearTimeout(timeout);
+  }, [navigate]);
+}
